changed decimal/double error, fixed minor issues

+++ HabiWei/Models/WeightEntry.cs
@@ -4,6 +4,6 @@
     {
         public int Id { get; set; }
         public DateTime Date { get; set; }
-        public double Weight { get; set; }
+        public decimal Weight { get; set; }
     }
 }
